+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * #25 Item-not-found-error caused by long term request.
 * #30 Base64 encoding problem in some older browsers.
 * #45 Move xhlr plugin to strophejs-plugins repo.
+* #62 Add `xmlunescape` method.
 * #70 `_onDisconnectTimeout` never tiggers because maxRetries is undefined.
 * #73 `getElementsByTagName` problem with namespaced elements.
 * #90 The queue data was not reset in .reset() method.